File structure and individual3 complete
