+++ src/css/style.css
@@ -109,14 +109,22 @@ button {
 
 @keyframes shake {
   10%,
-  90% {transform: translate3d(-1px, 0, 0);}
+  90% {
+    transform: translate3d(-5px, 0, 0);
+  }
   20%,
-  80% {transform: translate3d(2px, 0, 0);}
+  80% {
+    transform: translate3d(5px, 0, 0);
+  }
   30%,
   50%,
-  70% {transform: translate3d(-4px, 0, 0);}
+  70% {
+    transform: translate3d(-9px, 0, 0);
+  }
   40%,
-  60% {transform: translate3d(4px, 0, 0);}
+  60% {
+    transform: translate3d(9px, 0, 0);
+  }
 }
 
 .cart svg {
