+++ src/index.html
@@ -6,8 +6,8 @@
     <title>Sleep Outside | Home</title>
     <link rel="stylesheet" href="/css/style.css" />
     <script src="./js/main.js" type="module" defer></script>
-    <script src="./js/product.js" type="module"></script>
-    <script src="./js/productListing.mjs" type="module"></script>
+    <!--<script src="./js/product.js" type="module"></script>
+    <script src="./js/productListing.mjs" type="module"></script>-->
   </head>
 
   <body>
@@ -40,7 +40,8 @@
             <a href="product-listing/index.html?category=tents">
               <img
                 src="https://byui-cit.github.io/wdd330/images/category-icons/noun_Tent_3544767.svg"
-                alt="Tent Icon from Noun Project: Mustofa Bayu"/>
+                alt="Tent Icon from Noun Project: Mustofa Bayu"
+              />
               <h2>Tents</h2>
             </a>
           </li>
@@ -48,7 +49,8 @@
             <a href="product-listing/index.html?category=backpacks">
               <img
                 src="https://byui-cit.github.io/wdd330/images/category-icons/noun_carrier%20bag_3544771.svg"
-                alt="Backpack Icon from Noun Project: Mustofa Bayu"/>
+                alt="Backpack Icon from Noun Project: Mustofa Bayu"
+              />
               <h2>Backpacks</h2>
             </a>
           </li>
@@ -56,7 +58,8 @@
             <a href="product-listing/index.html?category=sleeping-bags">
               <img
                 src="https://byui-cit.github.io/wdd330/images/category-icons/noun_Sleeping%20Bag_3544775.svg"
-                alt="Sleeping Bag Icon from Noun Project: Mustofa Bayu"/>
+                alt="Sleeping Bag Icon from Noun Project: Mustofa Bayu"
+              />
               <h2>Sleeping Bags</h2>
             </a>
           </li>
@@ -64,12 +67,13 @@
             <a href="product-listing/index.html?category=hammocks">
               <img
                 src="https://byui-cit.github.io/wdd330/images/category-icons/noun_Hammock_791143.svg"
-                alt="Hammock Icon from Noun Project: Paul Richard"/>
+                alt="Hammock Icon from Noun Project: Paul Richard"
+              />
               <h2>Hammocks</h2>
             </a>
           </li>
         </ul>
-    </section>  
+      </section>
     </main>
     <footer id="footer"></footer>
   </body>
