+++ src/js/product.js
@@ -1,10 +1,8 @@
-import { getParams, loadHeaderFooter } from "./utils.mjs";
+import { getParams } from "./utils.mjs";
 import ProductDetails from "./ProductDetails.mjs";
 import ProductData from "./ProductData.mjs";
 
-loadHeaderFooter();
-
-const dataSource = new ProductData("tents");
+const dataSource = new ProductData();
 const productId = getParams("product");
 const product = new ProductDetails(productId, dataSource);
