+++ src/js/productListing.mjs
@@ -1,8 +1,7 @@
 import ProductData from "./ProductData.mjs";
 import productList from "./productList.mjs";
-import { loadHeaderFooter, getParams } from "./utils.mjs";
+import { getParams } from "./utils.mjs";
 
-loadHeaderFooter();
 
 const category = getParams("category");
 const dataSource = new ProductData();
