+++ src/product-listing/index.html
@@ -5,6 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Sleep Outside | Product-Listing</title>
     <link rel="stylesheet" href="../css/style.css" />
+    <script src="../js/main.js" type="module"></script>
     <script src="../js/productListing.mjs" type="module"></script>
   </head>
 
@@ -12,7 +13,7 @@
     <header class="divider" id="header"></header>
     <main class="divider">
       <section class="products">
-        <h2>Top Products <span class="title"></span></h2>
+        <h2>Top Products: <span class="title"></span></h2>
         <ul class="product-list"></ul>
       </section>
     </main>
