+++ src/product_pages/index.html
@@ -11,8 +11,7 @@
     <link rel="stylesheet" href="../css/style.css" />
     <script src="../js/main.js" type="module"></script>
     <script src="../js/product.js" type="module"></script>
-    <script src="../js/ProductDetails.mjs" type="module"></script>
-    <script src="../js/alert.js"></script>
+    <script src="../js/alert.mjs" type="module"></script>
   </head>
 
   <body>
